pyproject.toml
[project]
requires-python = ">=3.12"
name = "armada-den"
version = "0.1.0"
description = "Armada Den is an internal AI workspace built to enhance productivity, research, and communication within Hire Armada. It unifies AI chat, real-time knowledge access, and Gmail-based communication into one secure workspace, purpose-built for internal operations and client correspondence."
authors = [
    { name = "Natnael Alemseged", email = "natiaabaydam@gmail.com" },
]
dependencies = [
    "fastapi>=0.103.0",
    "fastapi-users[sqlalchemy] (>=15.0.1,<16.0.0)",
    "uvicorn==0.23.2",
    "pydantic>=2.0.0",
    "requests>=2.32.0",
    "loguru>=0.7.0",
    "joblib>=1.2.0",
    "scikit-learn>=1.1.3",
    "pandas>=2.2.3",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "composio>=0.9.1",
    "composio-openai>=0.9.1",
    "composio-agno>=0.7.20",
    "composio-llamaindex>=0.9.1",
    "llama-index>=0.10.0",
    "llama-index-llms-openai-like>=0.1.0",
    "llama-index-llms-groq>=0.1.0",
    "llama-index-core>=0.10.0",
    "openai>=1.0.0",
    "groq>=0.4.0",
    "agno>=0.1.0",
    "aiosqlite>=0.19.0",
    "authlib>=1.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "python-socketio>=5.11.0",
    "supabase>=2.3.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "supermemory>=0.2.0",
    "starlette>=0.27.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2",
    "black>=24.3",
    "autopep8>=2.0.0",
    "ipdb>=0.13.0",
    "pylint>=3.0.0"
]
aws = [
    "mangum>=0.17.0"
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]